apps/shipping/requirements.txt
fastapi==0.63.0
gunicorn==20.0.4
h11==0.12.0
pydantic==1.6.2
starlette==0.13.6
uvicorn==0.13.4
uvloop==0.16.0
websockets==8.1
httptools==0.1.2
python-dotenv==0.19.1
azure-cosmos==4.2.0
opencensus-ext-azure==1.1.0
opencensus-ext-requests==0.7.6
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
azure-monitor-opentelemetry-exporter
opentelemetry-instrumentation-requests
asgiref